Custom styling posts.
Actually got pretty far. The header and most of the post content is
done. A few notable exceptions are images and tables. The site footer
and index also need styling. Oh, and there needs to be mobile styling!
Of course. So yes. Essentially there is a lot still to do.